+++ service/api/auth_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/ant0ine/go-json-rest/rest"
 
 	testAuth "github.com/tidepool-org/platform/auth/test"
-	testErrors "github.com/tidepool-org/platform/errors/test"
+	errorsTest "github.com/tidepool-org/platform/errors/test"
 	"github.com/tidepool-org/platform/log"
 	logNull "github.com/tidepool-org/platform/log/null"
 	"github.com/tidepool-org/platform/request"
@@ -48,7 +48,7 @@ var _ = Describe("Auth", func() {
 			Expect(requireFunc).ToNot(BeNil())
 			requireFunc(res, req)
 			Expect(res.WriteHeaderInputs).To(BeEmpty())
-			Expect(res.WriteJsonInputs).To(BeEmpty())
+			Expect(res.WriteInputs).To(BeEmpty())
 		})
 
 		Context("with handlerFunc func", func() {
@@ -62,22 +62,21 @@ var _ = Describe("Auth", func() {
 			It("does nothing if response is nil", func() {
 				requireFunc(nil, req)
 				Expect(res.WriteHeaderInputs).To(BeEmpty())
-				Expect(res.WriteJsonInputs).To(BeEmpty())
+				Expect(res.WriteInputs).To(BeEmpty())
 			})
 
 			It("does nothing if request is nil", func() {
 				requireFunc(res, nil)
 				Expect(res.WriteHeaderInputs).To(BeEmpty())
-				Expect(res.WriteJsonInputs).To(BeEmpty())
+				Expect(res.WriteInputs).To(BeEmpty())
 			})
 
 			It("responds with unauthenticated error if details are missing", func() {
-				res.WriteJsonOutputs = []error{nil}
 				res.WriteOutputs = []testRest.WriteOutput{{BytesWritten: 0, Error: nil}}
 				requireFunc(res, req)
 				Expect(res.WriteHeaderInputs).To(Equal([]int{401}))
-				Expect(res.WriteJsonInputs).To(HaveLen(1))
-				testErrors.ExpectEqual(res.WriteJsonInputs[0].(error), request.ErrorUnauthenticated())
+				Expect(res.WriteInputs).To(HaveLen(1))
+				errorsTest.ExpectErrorJSON(request.ErrorUnauthenticated(), res.WriteInputs[0])
 			})
 
 			Context("with server details", func() {
@@ -88,7 +87,7 @@ var _ = Describe("Auth", func() {
 				It("responds successfully", func() {
 					requireFunc(res, req)
 					Expect(res.WriteHeaderInputs).To(Equal([]int{0}))
-					Expect(res.WriteJsonInputs).To(BeEmpty())
+					Expect(res.WriteInputs).To(BeEmpty())
 				})
 			})
 
@@ -100,7 +99,7 @@ var _ = Describe("Auth", func() {
 				It("responds successfully", func() {
 					requireFunc(res, req)
 					Expect(res.WriteHeaderInputs).To(Equal([]int{0}))
-					Expect(res.WriteJsonInputs).To(BeEmpty())
+					Expect(res.WriteInputs).To(BeEmpty())
 				})
 			})
 		})
@@ -112,7 +111,7 @@ var _ = Describe("Auth", func() {
 			Expect(requireFunc).ToNot(BeNil())
 			requireFunc(res, req)
 			Expect(res.WriteHeaderInputs).To(BeEmpty())
-			Expect(res.WriteJsonInputs).To(BeEmpty())
+			Expect(res.WriteInputs).To(BeEmpty())
 		})
 
 		Context("with handlerFunc func", func() {
@@ -126,22 +125,21 @@ var _ = Describe("Auth", func() {
 			It("does nothing if response is nil", func() {
 				requireFunc(nil, req)
 				Expect(res.WriteHeaderInputs).To(BeEmpty())
-				Expect(res.WriteJsonInputs).To(BeEmpty())
+				Expect(res.WriteInputs).To(BeEmpty())
 			})
 
 			It("does nothing if request is nil", func() {
 				requireFunc(res, nil)
 				Expect(res.WriteHeaderInputs).To(BeEmpty())
-				Expect(res.WriteJsonInputs).To(BeEmpty())
+				Expect(res.WriteInputs).To(BeEmpty())
 			})
 
 			It("responds with unauthenticated error if details are missing", func() {
-				res.WriteJsonOutputs = []error{nil}
 				res.WriteOutputs = []testRest.WriteOutput{{BytesWritten: 0, Error: nil}}
 				requireFunc(res, req)
 				Expect(res.WriteHeaderInputs).To(Equal([]int{401}))
-				Expect(res.WriteJsonInputs).To(HaveLen(1))
-				testErrors.ExpectEqual(res.WriteJsonInputs[0].(error), request.ErrorUnauthenticated())
+				Expect(res.WriteInputs).To(HaveLen(1))
+				errorsTest.ExpectErrorJSON(request.ErrorUnauthenticated(), res.WriteInputs[0])
 			})
 
 			Context("with server details", func() {
@@ -152,7 +150,7 @@ var _ = Describe("Auth", func() {
 				It("responds successfully", func() {
 					requireFunc(res, req)
 					Expect(res.WriteHeaderInputs).To(Equal([]int{0}))
-					Expect(res.WriteJsonInputs).To(BeEmpty())
+					Expect(res.WriteInputs).To(BeEmpty())
 				})
 			})
 
@@ -162,12 +160,11 @@ var _ = Describe("Auth", func() {
 				})
 
 				It("responds with unauthorized error", func() {
-					res.WriteJsonOutputs = []error{nil}
 					res.WriteOutputs = []testRest.WriteOutput{{BytesWritten: 0, Error: nil}}
 					requireFunc(res, req)
 					Expect(res.WriteHeaderInputs).To(Equal([]int{403}))
-					Expect(res.WriteJsonInputs).To(HaveLen(1))
-					testErrors.ExpectEqual(res.WriteJsonInputs[0].(error), request.ErrorUnauthorized())
+					Expect(res.WriteInputs).To(HaveLen(1))
+					errorsTest.ExpectErrorJSON(request.ErrorUnauthorized(), res.WriteInputs[0])
 				})
 			})
 		})
@@ -179,7 +176,7 @@ var _ = Describe("Auth", func() {
 			Expect(requireFunc).ToNot(BeNil())
 			requireFunc(res, req)
 			Expect(res.WriteHeaderInputs).To(BeEmpty())
-			Expect(res.WriteJsonInputs).To(BeEmpty())
+			Expect(res.WriteInputs).To(BeEmpty())
 		})
 
 		Context("with handlerFunc func", func() {
@@ -193,22 +190,21 @@ var _ = Describe("Auth", func() {
 			It("does nothing if response is nil", func() {
 				requireFunc(nil, req)
 				Expect(res.WriteHeaderInputs).To(BeEmpty())
-				Expect(res.WriteJsonInputs).To(BeEmpty())
+				Expect(res.WriteInputs).To(BeEmpty())
 			})
 
 			It("does nothing if request is nil", func() {
 				requireFunc(res, nil)
 				Expect(res.WriteHeaderInputs).To(BeEmpty())
-				Expect(res.WriteJsonInputs).To(BeEmpty())
+				Expect(res.WriteInputs).To(BeEmpty())
 			})
 
 			It("responds with unauthenticated error if details are missing", func() {
-				res.WriteJsonOutputs = []error{nil}
 				res.WriteOutputs = []testRest.WriteOutput{{BytesWritten: 0, Error: nil}}
 				requireFunc(res, req)
 				Expect(res.WriteHeaderInputs).To(Equal([]int{401}))
-				Expect(res.WriteJsonInputs).To(HaveLen(1))
-				testErrors.ExpectEqual(res.WriteJsonInputs[0].(error), request.ErrorUnauthenticated())
+				Expect(res.WriteInputs).To(HaveLen(1))
+				errorsTest.ExpectErrorJSON(request.ErrorUnauthenticated(), res.WriteInputs[0])
 			})
 
 			Context("with server details", func() {
@@ -217,12 +213,11 @@ var _ = Describe("Auth", func() {
 				})
 
 				It("responds with unauthorized error", func() {
-					res.WriteJsonOutputs = []error{nil}
 					res.WriteOutputs = []testRest.WriteOutput{{BytesWritten: 0, Error: nil}}
 					requireFunc(res, req)
 					Expect(res.WriteHeaderInputs).To(Equal([]int{403}))
-					Expect(res.WriteJsonInputs).To(HaveLen(1))
-					testErrors.ExpectEqual(res.WriteJsonInputs[0].(error), request.ErrorUnauthorized())
+					Expect(res.WriteInputs).To(HaveLen(1))
+					errorsTest.ExpectErrorJSON(request.ErrorUnauthorized(), res.WriteInputs[0])
 				})
 			})
 
@@ -234,7 +229,7 @@ var _ = Describe("Auth", func() {
 				It("responds successfully", func() {
 					requireFunc(res, req)
 					Expect(res.WriteHeaderInputs).To(Equal([]int{0}))
-					Expect(res.WriteJsonInputs).To(BeEmpty())
+					Expect(res.WriteInputs).To(BeEmpty())
 				})
 			})
 		})