+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+* Refactor request responder for more flexible responses; use direct http.ResponseWriter
 * Minor test refactor and cleanup
 * Add shortcuts for request error comparisons
 * Add streamed response from client in addition to existing data response